fix: emit radially-signed ring-edge normals
The absorbed-translation derivation assumes normals_vu carries a
consistent outward-radial sense.  Both models DOCUMENT that;
_polyline_from_edge_mask did not deliver it -- it fixes the normal axis
by a first-match neighbour test, so the sign followed scan order and
rasterization.  Measured on a rasterized closed ring through the shipped
code: mean|dot| with true radial 0.83 (axis right) but mean dot 0.003
(sign random), giving a sensitivity of 0.88 along the rasterizer's
quadrant bias where the true value is ~0.  The dilation and
opposite-side cancellations the derivation rests on could not occur on
any real frame.

Fixed at the source: the ring-radius backplane is already computed, so
each normal is oriented toward increasing radius.  Same measurement
after: mean dot 0.827 (signed == absolute) and sensitivity 0.0000.  The
docstring's false 'the +1 side is the high-radius side' claim is
corrected and a test pins both the sense and the closed-ring limit.

Also splits nav_model_rings.py (1154 lines) into ring_polyline and
ring_emission, and raises _median_radial_scale's unusable-scale sentinel
into a ValueError: a 1e-6 km/px floor would have produced a finite but
garbage ~1e12 px^2 covariance instead of failing.
